+++ docs/pages/documentation/general/GeneralNav.vue
@@ -1,0 +1,12 @@
+<template>
+    <div class="navbar-brand">
+        <router-link to="layout" class="navbar-item is-tab">Layout</router-link>
+        <router-link to="icon" class="navbar-item is-tab">Icon</router-link>
+        <router-link to="dropdown" class="navbar-item is-tab">Dropdown</router-link>
+        <router-link to="modal" class="navbar-item is-tab">Modal</router-link>
+        <router-link to="panel" class="navbar-item is-tab">Panel</router-link>
+        <router-link to="tabs" class="navbar-item is-tab">Tabs</router-link>
+        <router-link to="tag" class="navbar-item is-tab">Tag</router-link>
+        <router-link to="tooltip" class="navbar-item is-tab">Tooltip</router-link>
+    </div>
+</template>