+++ docs/pages/documentation/general/tabs/examples/ExPosition.vue
@@ -1,0 +1,15 @@
+<template>
+    <section>
+        <b-tabs position="is-centered" class="block">
+            <b-tab-item label="Pictures"></b-tab-item>
+            <b-tab-item label="Music"></b-tab-item>
+            <b-tab-item label="Videos"></b-tab-item>
+        </b-tabs>
+
+        <b-tabs position="is-right" class="block">
+            <b-tab-item label="Pictures"></b-tab-item>
+            <b-tab-item label="Music"></b-tab-item>
+            <b-tab-item label="Videos"></b-tab-item>
+        </b-tabs>
+    </section>
+</template>