+++ docs/pages/documentation/general/tag/examples/ExSimple.vue
@@ -1,0 +1,6 @@
+<template>
+    <section>
+        <b-tag>Tag label</b-tag>
+        <b-tag rounded>Rounded tag label</b-tag>
+    </section>
+</template>