+++ docs/pages/documentation/general/tag/examples/ExTaglist.vue
@@ -1,0 +1,9 @@
+<template>
+    <b-taglist>
+        <b-tag type="is-info">First</b-tag>
+        <b-tag type="is-info">Second</b-tag>
+        <b-tag type="is-info">Third</b-tag>
+        <b-tag type="is-info">Fourth</b-tag>
+        <b-tag type="is-info">Fifth</b-tag>
+    </b-taglist>
+</template>