+++ docs/pages/documentation/general/tag/examples/ExTaglistAttached.vue
@@ -1,0 +1,6 @@
+<template>
+    <b-taglist attached>
+        <b-tag type="is-dark">npm</b-tag>
+        <b-tag type="is-info">0.5.1</b-tag>
+    </b-taglist>
+</template>