+++ docs/pages/documentation/notices/Notification/examples/ExIcons.vue
@@ -1,0 +1,19 @@
+<template>
+    <section>
+        <b-notification type="is-info" has-icon>
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce id fermentum quam. Proin sagittis, nibh id hendrerit imperdiet, elit sapien laoreet elit
+        </b-notification>
+
+        <b-notification type="is-success" has-icon>
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce id fermentum quam. Proin sagittis, nibh id hendrerit imperdiet, elit sapien laoreet elit
+        </b-notification>
+
+        <b-notification type="is-warning" has-icon>
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce id fermentum quam. Proin sagittis, nibh id hendrerit imperdiet, elit sapien laoreet elit
+        </b-notification>
+
+        <b-notification type="is-danger" has-icon>
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce id fermentum quam. Proin sagittis, nibh id hendrerit imperdiet, elit sapien laoreet elit
+        </b-notification>
+    </section>
+</template>