+++ docs/router/index.js
@@ -7,29 +7,30 @@ const Home = () => import('../pages/Home')
 
 const Documentation = () => import('../pages/documentation/Documentation')
 
-import Installation from '../pages/documentation/installation/_Installation'
+import Installation from '../pages/documentation/installation/InstallationNav'
 const Start = () => import('../pages/documentation/installation/Start')
 const Customization = () => import('../pages/documentation/installation/Customization')
 const ConstructorOptions = () => import('../pages/documentation/installation/ConstructorOptions')
 
-import General from '../pages/documentation/general/_General'
-const Layout = () => import('../pages/documentation/general/Layout')
-const Icon = () => import('../pages/documentation/general/Icon')
-const Dropdown = () => import('../pages/documentation/general/Dropdown')
-const Modal = () => import('../pages/documentation/general/Modal')
-const Panel = () => import('../pages/documentation/general/Panel')
-const Tabs = () => import('../pages/documentation/general/Tabs')
-const Tooltip = () => import('../pages/documentation/general/Tooltip')
+import General from '../pages/documentation/general/GeneralNav'
+const Layout = () => import('../pages/documentation/general/layout')
+const Icon = () => import('../pages/documentation/general/icon')
+const Dropdown = () => import('../pages/documentation/general/dropdown')
+const Modal = () => import('../pages/documentation/general/modal')
+const Panel = () => import('../pages/documentation/general/panel')
+const Tabs = () => import('../pages/documentation/general/tabs')
+const Tag = () => import('../pages/documentation/general/tag')
+const Tooltip = () => import('../pages/documentation/general/tooltip')
 
-import Notices from '../pages/documentation/notices/_Notices'
-const Dialog = () => import('../pages/documentation/notices/Dialog')
-const Toast = () => import('../pages/documentation/notices/Toast')
-const Snackbar = () => import('../pages/documentation/notices/Snackbar')
-const Notification = () => import('../pages/documentation/notices/Notification')
-const Message = () => import('../pages/documentation/notices/Message')
-const Loading = () => import('../pages/documentation/notices/Loading')
+import Notices from '../pages/documentation/notices/NoticesNav'
+const Dialog = () => import('../pages/documentation/notices/dialog')
+const Toast = () => import('../pages/documentation/notices/toast')
+const Snackbar = () => import('../pages/documentation/notices/snackbar')
+const Notification = () => import('../pages/documentation/notices/notification')
+const Message = () => import('../pages/documentation/notices/message')
+const Loading = () => import('../pages/documentation/notices/loading')
 
-import FormElements from '../pages/documentation/form/_Form'
+import FormElements from '../pages/documentation/form/FormNav'
 const Field = () => import('../pages/documentation/form/Field')
 const Input = () => import('../pages/documentation/form/Input')
 const Select = () => import('../pages/documentation/form/Select')
@@ -40,9 +41,9 @@ const Switch = () => import('../pages/documentation/form/Switch')
 const Upload = () => import('../pages/documentation/form/Upload')
 const Datepicker = () => import('../pages/documentation/form/Datepicker')
 
-import DataManipulation from '../pages/documentation/data/_Data'
-const Table = () => import('../pages/documentation/data/Table')
-const Pagination = () => import('../pages/documentation/data/Pagination')
+import DataManipulation from '../pages/documentation/data/DataNav'
+const Table = () => import('../pages/documentation/data/table')
+const Pagination = () => import('../pages/documentation/data/pagination')
 
 export default new Router({
     mode: 'hash',
@@ -123,6 +124,13 @@ export default new Router({
                 {
                     path: 'panel',
                     component: Panel,
+                    meta: {
+                        category: General
+                    }
+                },
+                {
+                    path: 'tag',
+                    component: Tag,
                     meta: {
                         category: General
                     }