+++ src/components/tag/Taglist.vue
@@ -1,0 +1,14 @@
+<template>
+    <div class="tags" :class="{ 'has-addons': attached }">
+        <slot></slot>
+    </div>
+</template>
+
+<script>
+    export default {
+        name: 'bTaglist',
+        props: {
+            attached: Boolean
+        }
+    }
+</script>